refactor: send 관련 클래스들을 별도 패키지로 묶어서 정리
