+++ src/test/java/umc/cockple/demo/domain/chat/service/websocket/send/support/ChatMessageFileAppenderTest.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.domain.chat.service.support;
+package umc.cockple.demo.domain.chat.service.websocket.send.support;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
