+++ src/test/java/umc/cockple/demo/domain/chat/service/websocket/send/support/DirectChatRoomActivationServiceTest.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.domain.chat.service.support;
+package umc.cockple.demo.domain.chat.service.websocket.send.support;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
